Move some functions from lib ro rc-utils
