+++ packages/rocketchat-lib/lib/getURL.js
@@ -1,21 +1,3 @@
-import { Meteor } from 'meteor/meteor';
-import s from 'underscore.string';
+import { getURL } from 'meteor/rocketchat:utils';
 
-RocketChat.getURL = (path, { cdn = true, full = false } = {}) => {
-	const cdnPrefix = s.rtrim(s.trim(RocketChat.settings.get('CDN_PREFIX') || ''), '/');
-	const pathPrefix = s.rtrim(s.trim(__meteor_runtime_config__.ROOT_URL_PATH_PREFIX || ''), '/');
-
-	let basePath;
-
-	const finalPath = s.ltrim(s.trim(path), '/');
-
-	if (cdn && cdnPrefix !== '') {
-		basePath = cdnPrefix + pathPrefix;
-	} else if (full) {
-		return Meteor.absoluteUrl(finalPath);
-	} else {
-		basePath = pathPrefix;
-	}
-
-	return `${ basePath }/${ finalPath }`;
-};
+RocketChat.getURL = getURL;
