+++ packages/rocketchat-utils/server/index.js
@@ -1,32 +1,16 @@
-import { t, isRtl } from '../lib/tapi18n';
-import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
-import { Info } from '../rocketchat.info';
-import { getUserPreference } from '../lib/getUserPreference';
-import { fileUploadMediaWhiteList, fileUploadIsValidContentType } from '../lib/fileUploadRestrictions';
-import { roomTypes } from './lib/roomTypes';
-import { RoomTypeRouteConfig, RoomTypeConfig, RoomSettingsEnum, UiTextContext } from '../lib/RoomTypeConfig';
-import { RoomTypesCommon } from '../lib/RoomTypesCommon';
-import { isDocker } from './functions/isDocker';
-import { getAvatarUrlFromUsername } from '../lib/getAvatarUrlFromUsername';
-import { slashCommands } from '../lib/slashCommand';
-import { getUserNotificationPreference } from '../lib/getUserNotificationPreference';
-
-export {
-	t,
-	isRtl,
-	getDefaultSubscriptionPref,
-	Info,
-	getUserPreference,
-	fileUploadIsValidContentType,
-	fileUploadMediaWhiteList,
-	roomTypes,
-	RoomTypeRouteConfig,
-	RoomTypesCommon,
-	RoomTypeConfig,
-	RoomSettingsEnum,
-	UiTextContext,
-	isDocker,
-	getAvatarUrlFromUsername,
-	slashCommands,
-	getUserNotificationPreference,
-};
+export { t, isRtl } from '../lib/tapi18n';
+export { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
+export { Info } from '../rocketchat.info';
+export { getUserPreference } from '../lib/getUserPreference';
+export { fileUploadMediaWhiteList, fileUploadIsValidContentType } from '../lib/fileUploadRestrictions';
+export { roomTypes } from './lib/roomTypes';
+export { RoomTypeRouteConfig, RoomTypeConfig, RoomSettingsEnum, UiTextContext } from '../lib/RoomTypeConfig';
+export { RoomTypesCommon } from '../lib/RoomTypesCommon';
+export { isDocker } from './functions/isDocker';
+export { getAvatarUrlFromUsername } from '../lib/getAvatarUrlFromUsername';
+export { slashCommands } from '../lib/slashCommand';
+export { getUserNotificationPreference } from '../lib/getUserNotificationPreference';
+export { getAvatarColor } from '../lib/getAvatarColor';
+export { getURL } from '../lib/getURL';
+export { getValidRoomName } from '../lib/getValidRoomName';
+export { placeholders } from '../lib/placeholders';
